Fix cleanup code, leaderboard scraper, dockerfile
There have been more weird json choices that make my cleanup code not work for every situation, so fix it to work and try to generalize it further again. The leaderboard was updated again and the data was moved; readjust. And docker needed to be adjusted slightly for Python PEP668.